small fixes CPP07 -> all exercises

+++ CPP07/ex01/include/iter.hpp
@@ -9,7 +9,7 @@
     {
         std::cout << "The address of the array: " << &arr << " | ";
         std::cout << "The length of the array: " << arr_len << " | " << std::endl;
-        for (size_t i = 0; i < arr_len; i++)
+        for (int i = 0; i < arr_len; i++)
         {
             (*function)(arr[i]);
         }
